Extract browser DOM snapshot primitive
